remove not used beehive common package
